% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkfiles.R
\name{checkfiles}
\alias{checkfiles}
\title{Checks the min temperature, max tempature and precipitation files in the working directory and whether there are any expected errors.}
\usage{
checkfiles()
}
\value{
test results
}
\description{
Checks the min temperature, max tempature and precipitation files in the working directory and whether there are any expected errors.
}
\examples{
checkfiles()
}
