% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepclmdata.R
\name{prepclmdata}
\alias{prepclmdata}
\title{Prepares the CLM data}
\usage{
prepclmdata(xmin = -180, xmax = 180, ymin = -90, ymax = 90)
}
\arguments{
\item{xmin}{westmost longitudinal extent (in decimal degrees). Precede with minus if necessary. Default: -180째}

\item{xmax}{eastmost longitudinal extent (in decimal degrees). Precede with minus if necessary. Default: 180째}

\item{ymin}{southmost latitudinal extent (in decimal degrees). Precede with minus if necessary. Default: -90째}

\item{ymax}{northmost latitudinal extent (in decimal degrees). Precede with minus if necessary. Default: 90째}
}
\value{
prepared .grd files for DIVA-GIS
}
\description{
Prepares the CLM data for import into DIVA-GIS.

BEFORE USING: ensure that you have either: 1) set your working directory and ran prepexample(), or 2) have downloaded the necessary data in your working directory:
\itemize{
\item \strong{minimum temperature data} - often labelled tn
\item \strong{average temperature data} -  often labelled tx
\item \strong{precipitation data} - often labelled pr
}

IMPORTANT: Make note of the Directory once the instruction completes - you will need to navigate to it to create the CLM file in DIVA-GIS.
}
\examples{
#prepare CLM data for global coverage. NOTE: if using high resolution data, global coverage this can use a VAST amount of memory and storage space. We recommend the second example if this is the case.
prepclmdata()

#prepare CLM data for around London/SE England
prepclmdata(-1, 1, 50, 52)
}
