\name{prepclmdata}
\alias{prepclmdata}
\title{Prepares the CLM data}
\usage{
prepclmdata()
}
\description{
Prepares the CLM data for import into DIVA-GIS.

BEFORE USING: ensure that you have either: 1) set your working directory and ran prepexample(), or 2) have downloaded the necessary data in your working directory:
- \strong{minimum temperature data} - often labelled tn
- \strong{average temperature data} -  often labelled tx
- \strong{precipitation data} - often labelled pr

IMPORTANT: Make note of the Directory once the instruction completes - you will need to navigate to it to create the CLM file in DIVA-GIS.
}
\examples{
prepclmdata()
}

